Revert "Merge branch 'main' of https://github.com/ChefAtHands/utility-services"
This reverts commit cda3a0f58e96efd7f951c00ea709818c2d5f53fd, reversing
changes made to 9fe78dbb7ec42e14c68465e072576a1e4f84a83f.

+++ ingredient-service/src/main/java/com/chefathands/ingredient/controller/IngredientController.java
@@ -20,7 +20,7 @@ public class IngredientController {
 
     //search by id
     @GetMapping("/{id}")
-    public ResponseEntity<Ingredient> getById(@PathVariable Integer id) {
+    public ResponseEntity<Ingredient> getById(@PathVariable Long id) {
         return ResponseEntity.ok(service.findById(id));
     }
 
@@ -40,7 +40,7 @@ public class IngredientController {
 
     //delete ingredient by id
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
         service.findById(id); // Check if exists
         service.deleteById(id);
         return ResponseEntity.noContent().build();
