+++ ingredient-service/src/main/java/com/chefathands/ingredient/service/IngredientService.java
@@ -20,7 +20,7 @@ public class IngredientService {
         return repo.save(i); 
     }
     
-      public Ingredient findById(Integer id) {
+      public Ingredient findById(Long id) {
         return repo.findById(id)
             .orElseThrow(() -> new ResourceNotFoundException("Ingredient not found with id " + id));
     }
@@ -33,7 +33,7 @@ public class IngredientService {
         return repo.findByNameContainingIgnoreCase(name);
     }
 
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         repo.deleteById(id);
     }
